LayoutTests/imported/w3c/web-platform-tests/intersection-observer/svg-container-element-expected.txt


PASS IntersectionObserver observing an SVG container element changing position
FAIL First rAF. assert_approx_equals: entries[0].boundingClientRect.right expected 108 +/- 1 but got 8
FAIL Change inner svg element assert_approx_equals: entries[1].boundingClientRect.right expected 108 +/- 1 but got 8
